Moved and renamed VehicleLib to NotifyLib
